requirements.txt
pyyaml>=6.0
requests>=2.31
fastapi>=0.111
uvicorn>=0.29
python-multipart>=0.0.9
